Split out hooks and patches, passive disadv fn
