+++ scripts/patches.js
@@ -1,0 +1,43 @@
+import { Stealthy } from "./stealthy.js";
+
+Hooks.once('setup', () => {
+  libWrapper.register(
+    'stealthy',
+    'DetectionModeBasicSight.prototype.testVisibility',
+    (wrapped, visionSource, mode, config = {}) => {
+      if (!Stealthy.testVisionStealth(visionSource, config)) return false;
+
+      const target = config.object?.actor;
+      let noDarkvision = false;
+      const ignoreFriendlyUmbralSight =
+        game.settings.get('stealthy', 'ignoreFriendlyUmbralSight') &&
+        config.object.document?.disposition === visionSource.object.document?.disposition;
+      if (!ignoreFriendlyUmbralSight && visionSource.visionMode?.id === 'darkvision') {
+        const umbralSight = target?.itemTypes?.feat?.find(f => f.name === game.i18n.localize('Umbral Sight'));
+        if (umbralSight) noDarkvision = true;
+      }
+
+      if (noDarkvision) {
+        Stealthy.log(`${visionSource.object.name}'s darkvision can't see ${config.object.name}`);
+        let ourMode = duplicate(mode);
+        ourMode.range = 0;
+        return wrapped(visionSource, ourMode, config);
+      }
+
+      return wrapped(visionSource, mode, config);
+    },
+    libWrapper.MIXED,
+    { perf_mode: libWrapper.PERF_FAST }
+  );
+
+  libWrapper.register(
+    'stealthy',
+    'DetectionModeInvisibility.prototype.testVisibility',
+    (wrapped, visionSource, mode, config = {}) => {
+      if (!Stealthy.testVisionStealth(visionSource, config)) return false;
+      return wrapped(visionSource, mode, config);
+    },
+    libWrapper.MIXED,
+    { perf_mode: libWrapper.PERF_FAST }
+  );
+});
